requirements.txt
requests
types-requests
python-dotenv
pydantic